remove monitor task completely
